Fixed a bug in IsEqualTo
Apparently, when the `auto` keyword was added to the Go language in
support of generics, it broke the behavior of `reflect.DeepEqual`.
According to the comment, integers should be compared as if the `==`
operator was used. This no longer appears to be the case. Whether the
change in functionality was intentional and the documentation is
out-of-date, or the functionality broke compatibility with previous
versions of Go remains to be determined. There is an issue open
regarding this: https://github.com/golang/go/issues/56991